requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
click==7.1.2
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==3.1
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.15.0
regex==2020.11.13
rfc3986==1.4.0
sniffio==1.2.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
